refactor(desktop): co-locate browser pane integration, remove keep-alive
- Float-over webviews already preserve state across tab switches via the
  registry's attach/detach pair, so keep-alive tab rendering was redundant.
  Revert packages/panes Workspace.tsx to rendering only the active tab.
- Move all browser-specific wiring from page.tsx into a new
  useBrowserPaneIntegration hook co-located with BrowserPane. page.tsx now
  consumes a single hook call instead of owning tab title derivation, tab
  label rendering, and the destroy diff effect.
- Registry attach takes over the visibility flip directly; setVisibility
  method is no longer needed. Drop the rAF deferral — React useEffect fires
  after layout commit so getBoundingClientRect reads correct values
  synchronously.
- BrowserPane no longer subscribes to the workspace store for isTabActive.
