+++ apps/desktop/src/renderer/routes/_authenticated/_dashboard/v2-workspace/$workspaceId/hooks/usePaneRegistry/components/BrowserPane/index.ts
@@ -1,1 +1,2 @@
 export { BrowserPane, BrowserPaneToolbar } from "./BrowserPane";
+export { useBrowserPaneIntegration } from "./useBrowserPaneIntegration";
