+++ packages/panes/src/react/components/Workspace/Workspace.tsx
@@ -1,5 +1,4 @@
 import { cn } from "@superset/ui/utils";
-import { useEffect, useState } from "react";
 import { useStore } from "zustand";
 import type { WorkspaceProps } from "../../types";
 import { Tab } from "./components/Tab";
@@ -21,36 +20,7 @@ export function Workspace<TData>({
 }: WorkspaceProps<TData>) {
 	const tabs = useStore(store, (s) => s.tabs);
 	const activeTabId = useStore(store, (s) => s.activeTabId);
-
-	const [mountedTabIds, setMountedTabIds] = useState<Set<string>>(
-		() => new Set(activeTabId ? [activeTabId] : []),
-	);
-
-	useEffect(() => {
-		if (!activeTabId) return;
-		setMountedTabIds((prev) => {
-			if (prev.has(activeTabId)) return prev;
-			const next = new Set(prev);
-			next.add(activeTabId);
-			return next;
-		});
-	}, [activeTabId]);
-
-	useEffect(() => {
-		setMountedTabIds((prev) => {
-			const existing = new Set(tabs.map((t) => t.id));
-			let changed = false;
-			const next = new Set<string>();
-			for (const id of prev) {
-				if (existing.has(id)) {
-					next.add(id);
-				} else {
-					changed = true;
-				}
-			}
-			return changed ? next : prev;
-		});
-	}, [tabs]);
+	const activeTab = tabs.find((t) => t.id === activeTabId) ?? null;
 
 	const closeTab = async (tabId: string) => {
 		if (onBeforeCloseTab) {
@@ -96,36 +66,19 @@ export function Workspace<TData>({
 				renderTabAccessory={renderTabAccessory}
 			/>
 			{renderBelowTabBar?.()}
-			<div className="relative flex min-h-0 min-w-0 flex-1 overflow-hidden">
-				{tabs
-					.filter((tab) => mountedTabIds.has(tab.id))
-					.map((tab) => {
-						const isActive = tab.id === activeTabId;
-						return (
-							<div
-								key={tab.id}
-								className={cn(
-									"absolute inset-0 flex min-h-0 min-w-0 flex-col",
-									!isActive && "pointer-events-none invisible",
-								)}
-								aria-hidden={!isActive}
-							>
-								<Tab
-									store={store}
-									tab={tab}
-									registry={registry}
-									paneActions={paneActions}
-									contextMenuActions={contextMenuActions}
-								/>
-							</div>
-						);
-					})}
-				{mountedTabIds.size === 0 && (
-					<div className="flex min-h-0 min-w-0 flex-1 items-center justify-center text-sm text-muted-foreground">
-						{renderEmptyState?.() ?? "No tabs open"}
-					</div>
-				)}
-			</div>
+			{activeTab ? (
+				<Tab
+					store={store}
+					tab={activeTab}
+					registry={registry}
+					paneActions={paneActions}
+					contextMenuActions={contextMenuActions}
+				/>
+			) : (
+				<div className="flex min-h-0 min-w-0 flex-1 items-center justify-center text-sm text-muted-foreground">
+					{renderEmptyState?.() ?? "No tabs open"}
+				</div>
+			)}
 		</div>
 	);
 }
